Hotfix ETP-3915: Add unit and integration tests for coverage

+++ src-test/src/org/openbravo/advpaymentmngt/utility/FINUtilityTest.java
@@ -165,28 +165,28 @@ public class FINUtilityTest extends BaseCoreTest {
     @Test
     void testSameDate() {
       Date now = new Date();
-      assertEquals(0L, FIN_Utility.getDaysBetween(now, now));
+      assertEquals(0L, (long) FIN_Utility.getDaysBetween(now, now));
     }
 
     @Test
     void testOneDayApart() {
       Date begin = DateUtils.truncate(new Date(), Calendar.DATE);
       Date end = DateUtils.addDays(begin, 1);
-      assertEquals(1L, FIN_Utility.getDaysBetween(begin, end));
+      assertEquals(1L, (long) FIN_Utility.getDaysBetween(begin, end));
     }
 
     @Test
     void testNegativeDays() {
       Date begin = DateUtils.truncate(new Date(), Calendar.DATE);
       Date end = DateUtils.addDays(begin, -3);
-      assertEquals(-3L, FIN_Utility.getDaysBetween(begin, end));
+      assertEquals(-3L, (long) FIN_Utility.getDaysBetween(begin, end));
     }
 
     @Test
     void testTenDays() {
       Date begin = DateUtils.truncate(new Date(), Calendar.DATE);
       Date end = DateUtils.addDays(begin, 10);
-      assertEquals(10L, FIN_Utility.getDaysBetween(begin, end));
+      assertEquals(10L, (long) FIN_Utility.getDaysBetween(begin, end));
     }
   }
